% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brawand.htseq.fpkm.pooled}
\alias{brawand.htseq.fpkm.pooled}
\title{Gene expression for Brawand et al. (ArrayExpress E-GEOD-30352) mapped to Ensembl 76 (GRCh38)
and quantified with HTSeq-count and normalised with the FPKM method.
Then all the biological replicates have been pooled together (median values)
See for more details chapters 1 and 2 of Barzine, M.P PhD thesis:
Investigating Normal Human Gene Expression in Tissues
with High-throughput Transcriptomic and Proteomic data.}
\format{A data frame with 44479 transcritps (ie rows/observations) for 8 Tissues (ie columns/variables)
\describe{
\item{Frontalcortex}{numeric, FPKM}
\item{Prefrontal cortex}{numeric, FPKM}
\item{Temporal lobe}{numeric, FPKM}
\item{Cerebellum}{numeric, FPKM}
\item{Heart}{numeric, FPKM}
\item{Kidney}{numeric, FPKM}
\item{Liver}{numeric, FPKM}
\item{Testis}{numeric, FPKM}
}}
\usage{
brawand.htseq.fpkm.pooled
}
\description{
Gene expression for Brawand et al. (ArrayExpress E-GEOD-30352) mapped to Ensembl 76 (GRCh38)
and quantified with HTSeq-count and normalised with the FPKM method.
Then all the biological replicates have been pooled together (median values)
See for more details chapters 1 and 2 of Barzine, M.P PhD thesis:
Investigating Normal Human Gene Expression in Tissues
with High-throughput Transcriptomic and Proteomic data.
}
\keyword{datasets}
