% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brawand.htseq.fpkm}
\alias{brawand.htseq.fpkm}
\title{Gene expression for Brawand et al. (ArrayExpress E-GEOD-30352) mapped to Ensembl 76 (GRCh38)
and quantified with HTSeq-count and normalised with the FPKM method.
See for more details chapters 1 and 2 of Barzine, M.P PhD thesis:
Investigating Normal Human Gene Expression in Tissues
with High-throughput Transcriptomic and Proteomic data.}
\format{A data frame with 44479 transcritps (ie rows/observations) for 21 samples (ie columns/variables)
\describe{
\item{Prefrontal1}{numeric, FPKM}
\item{Prefrontal2}{numeric, FPKM}
\item{Frontal1}{numeric, FPKM}
\item{Frontal2}{numeric, FPKM}
\item{Prefrontal3}{numeric, FPKM}
\item{Temporal1}{numeric, FPKM}
\item{Cerebellum1}{numeric, FPKM}
\item{Cerebellum2}{numeric, FPKM}
\item{Cerebellum3}{numeric, FPKM}
\item{Heart1}{numeric, FPKM}
\item{Heart2}{numeric, FPKM}
\item{Heart3}{numeric, FPKM}
\item{Heart4}{numeric, FPKM}
\item{Kidney1}{numeric, FPKM}
\item{Kidney2}{numeric, FPKM}
\item{Kidney3}{numeric, FPKM}
\item{Liver1}{numeric, FPKM}
\item{Liver2}{numeric, FPKM}
\item{Liver3}{numeric, FPKM}
\item{Testis1}{numeric, FPKM}
\item{Testis2}{numeric, FPKM}
}}
\usage{
brawand.htseq.fpkm

brawand.htseq.fpkm
}
\description{
Gene expression for Brawand et al. (ArrayExpress E-GEOD-30352) mapped to Ensembl 76 (GRCh38)
and quantified with HTSeq-count and normalised with the FPKM method.
See for more details chapters 1 and 2 of Barzine, M.P PhD thesis:
Investigating Normal Human Gene Expression in Tissues
with High-throughput Transcriptomic and Proteomic data.

Gene expression for Brawand et al. (ArrayExpress E-GEOD-30352) mapped to Ensembl 76 (GRCh38)
and quantified with HTSeq-count.
See for more details chapters 1 and 2 of Barzine, M.P PhD thesis:
Investigating Normal Human Gene Expression in Tissues
with High-throughput Transcriptomic and Proteomic data.
}
\keyword{datasets}
