% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brawand.htseq.raw}
\alias{brawand.htseq.raw}
\title{Gene expression for Brawand et al. (ArrayExpress E-GEOD-30352) mapped to Ensembl 76 (GRCh38)
and quantified with HTSeq-count.
See for more details chapters 1 and 2 of Barzine, M.P PhD thesis:
Investigating Normal Human Gene Expression in Tissues
with High-throughput Transcriptomic and Proteomic data.}
\format{A data frame with 44479 transcritps (ie rows/observations) for 21 samples (ie columns/variables)
\describe{
\item{Prefrontal1}{integer, count}
\item{Prefrontal2}{integer, count}
\item{Frontal1}{integer, count}
\item{Frontal2}{integer, count}
\item{Prefrontal3}{integer, count}
\item{Temporal1}{integer, count}
\item{Cerebellum1}{integer, count}
\item{Cerebellum2}{integer, count}
\item{Cerebellum3}{integer, count}
\item{Heart1}{integer, count}
\item{Heart2}{integer, count}
\item{Heart3}{integer, count}
\item{Heart4}{integer, count}
\item{Kidney1}{integer, count}
\item{Kidney2}{integer, count}
\item{Kidney3}{integer, count}
\item{Liver1}{integer, count}
\item{Liver2}{integer, count}
\item{Liver3}{integer, count}
\item{Testis1}{integer, count}
\item{Testis2}{integer, count}
}}
\usage{
brawand.htseq.raw
}
\description{
Gene expression for Brawand et al. (ArrayExpress E-GEOD-30352) mapped to Ensembl 76 (GRCh38)
and quantified with HTSeq-count.
See for more details chapters 1 and 2 of Barzine, M.P PhD thesis:
Investigating Normal Human Gene Expression in Tissues
with High-throughput Transcriptomic and Proteomic data.
}
\keyword{datasets}
