% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{castle.htseq.fpkm}
\alias{castle.htseq.fpkm}
\title{Gene expression for Castle et al. (ArrayExpress E-MTAB-305) mapped to Ensembl 76 (GRCh38)
and quantified with HTSeq-count and normalised with the FPKM method.
See for more details chapters 1 and 2 of Barzine, M.P PhD thesis:
Investigating Normal Human Gene Expression in Tissues
with High-throughput Transcriptomic and Proteomic data.}
\format{A data frame with 43921 transcritps (ie rows/observations) for 11 tissues (ie columns/variables)
\describe{
\item{Adipose}{numeric, FPKM}
\item{Colon}{numeric, FPKM}
\item{Heart}{numeric, FPKM}
\item{Hypothalamus}{numeric, FPKM}
\item{Kidney}{numeric, FPKM}
\item{Liver}{numeric, FPKM}
\item{Lung}{numeric, FPKM}
\item{Ovary}{numeric, FPKM}
\item{Skeletal muscle}{numeric, FPKM}
\item{Spleen}{numeric, FPKM}
\item{Testis}{numeric, FPKM}
}}
\usage{
castle.htseq.fpkm
}
\description{
Gene expression for Castle et al. (ArrayExpress E-MTAB-305) mapped to Ensembl 76 (GRCh38)
and quantified with HTSeq-count and normalised with the FPKM method.
See for more details chapters 1 and 2 of Barzine, M.P PhD thesis:
Investigating Normal Human Gene Expression in Tissues
with High-throughput Transcriptomic and Proteomic data.
}
\keyword{datasets}
