% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{completePalette}
\alias{completePalette}
\title{Colour palette covering all the tissues of the five transcriptomic and three proteomics datasets
with their different spelling
designed with
[https://medialab.github.io/iwanthue/](https://medialab.github.io/iwanthue/)}
\format{A name vector with tissues as names and content the hex code for the colour}
\usage{
completePalette
}
\description{
Colour palette covering all the tissues of the five transcriptomic and three proteomics datasets
with their different spelling
designed with
[https://medialab.github.io/iwanthue/](https://medialab.github.io/iwanthue/)
}
\keyword{datasets}
