% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cutler.top3}
\alias{cutler.top3}
\title{Protein expression for Cutler Lab data mapped to Ensembl 76 (GRCh38)
and quantified with the Top3 method (PSM) explained
in the second chapter of Barzine, M.P PhD thesis:
Investigating Normal Human Gene Expression in Tissues
with High-throughput Transcriptomic and Proteomic data.}
\format{A data frame with 3351 proteins (ie rows/observations) for 10 tissue samples (ie columns/variables)
\describe{
\item{CSF}{numeric,normalised PSM}
\item{Adipose}{numeric,normalised PSM}
\item{Bone}{numeric,normalised PSM}
\item{Breast}{numeric,normalised PSM}
\item{Heart}{numeric,normalised PSM}
\item{Lung}{numeric,normalised PSM}
\item{Ovary}{numeric,normalised PSM}
\item{Pancreas}{numeric,normalised PSM}
\item{Platelets lysate}{numeric,normalised PSM}
\item{Platelets}{numeric,normalised PSM}
}}
\usage{
cutler.top3
}
\description{
Protein expression for Cutler Lab data mapped to Ensembl 76 (GRCh38)
and quantified with the Top3 method (PSM) explained
in the second chapter of Barzine, M.P PhD thesis:
Investigating Normal Human Gene Expression in Tissues
with High-throughput Transcriptomic and Proteomic data.
}
\keyword{datasets}
