% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{datasetCol}
\alias{datasetCol}
\title{Colour palette which associates one colour to each of the datasets,
proteomics or transcriptomics,
designed with
[https://medialab.github.io/iwanthue/](https://medialab.github.io/iwanthue/)}
\format{A name vector with tissues as names and content the hex code for the colour}
\usage{
datasetCol
}
\description{
Colour palette which associates one colour to each of the datasets,
proteomics or transcriptomics,
designed with
[https://medialab.github.io/iwanthue/](https://medialab.github.io/iwanthue/)
}
\keyword{datasets}
