% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{fixNamesBrawand}
\alias{fixNamesBrawand}
\title{Homogenise the names of the Brawand et al. data with the other datasets}
\usage{
fixNamesBrawand(DF)
}
\arguments{
\item{DF}{a Uhlen expression (numeric) data.frame}
}
\value{
a data.frame with homogenised columns names
}
\description{
Homogenise the names of the Brawand et al. data with the other datasets
}
