% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{fixNamesCutler}
\alias{fixNamesCutler}
\title{Homogenise the names of the Cutler data with the other datasets
(in particular with Uhlen et al. )}
\usage{
fixNamesCutler(DF)
}
\arguments{
\item{DF}{cutler expression (Top3 PSM) (numeric) data.frame}
}
\value{
a data.frame with homogenised columns names
}
\description{
Homogenise the names of the Cutler data with the other datasets
(in particular with Uhlen et al. )
}
