% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{fixNamesKuster}
\alias{fixNamesKuster}
\title{Homogenise the names of the Kuster data with the other datasets
(in particular with Uhlen et al. )}
\usage{
fixNamesKuster(DF)
}
\arguments{
\item{DF}{kuster expression (Top3 PSM) (numeric) data.frame}
}
\value{
a data.frame with homogenised columns names
}
\description{
Homogenise the names of the Kuster data with the other datasets
(in particular with Uhlen et al. )
}
