% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{fixNamesPandey}
\alias{fixNamesPandey}
\title{Homogenise the names of the Pandey data with the other datasets
(in particular with Uhlen et al. )}
\usage{
fixNamesPandey(DF, method)
}
\arguments{
\item{DF}{pandey expression (Top3 PSM) (numeric) data.frame}

\item{method}{character string. Allows to adapt to the original files.}
}
\value{
a data.frame with homogenised columns names
}
\description{
Homogenise the names of the Pandey data with the other datasets
(in particular with Uhlen et al. )
}
