% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{g.mitoch}
\alias{g.mitoch}
\title{Vector with ensembl id of genes that are annotated as found in the mitochondrion in Ensembl}
\format{A vector of character strings}
\usage{
g.mitoch
}
\description{
Vector with ensembl id of genes that are annotated as found in the mitochondrion in Ensembl
}
\keyword{datasets}
