% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{g.Pcoding}
\alias{g.Pcoding}
\title{Vector with ensembl id of genes that are annotated as coding in Ensembl 76}
\format{A vector of character strings}
\usage{
g.Pcoding
}
\description{
Vector with ensembl id of genes that are annotated as coding in Ensembl 76
}
\keyword{datasets}
