% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gene.mapID}
\alias{gene.mapID}
\title{Map between the Ensembl gene ids and their gene symbol
based on Ensembl 76 (GRCh38) and created with nameID()}
\format{A name vector with Ensembl gene ids as names and gene symbols (hgcn) as content}
\usage{
gene.mapID
}
\description{
Map between the Ensembl gene ids and their gene symbol
based on Ensembl 76 (GRCh38) and created with nameID()
}
\keyword{datasets}
