% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geneInfoEns76}
\alias{geneInfoEns76}
\title{Different gene information based on Ensembl 76 (GRCh38)}
\format{A data frame with 62753 rows and 11 variables:
\describe{
  \item{ensembl_gene_id}{gene ids in Ensembl 76 (also the row.names)}
  \item{gene_name}{hgcn symbol}
  \item{p_coding_biotype}{TRUE if the gene is annotated as protein coding in Ensembl 76}
  \item{evidence}{if the gene has been found as a protein in human protein atlas (www.https://www.proteinatlas.org/)}
  \item{description}{Gene associated description}
  \item{entrez_id}{collated list of corresponding entrez ids}
  \item{go_id}{collated list of corresponding GO ids}
  \item{go_desc}{collated list of description of the GO ids}
  \item{transcript_count}{number of transcripts annotated in Ensembl 76 for each gene}
  \item{nb.exons}{number of possible exons for each gene}
  \item{Gene.length}{considered length for each gene (length of the sum of the collated exons)}
}}
\source{
from several biomaRt and \url{http://www.ensembl.org/biomart} queries on Ensembl 76
which can not be done anymore on that version specifically
}
\usage{
geneInfoEns76
}
\description{
Different gene information based on Ensembl 76 (GRCh38)
}
\keyword{datasets}
