% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{goAnnot.76}
\alias{goAnnot.76}
\title{Different gene information based on Ensembl 76 (GRCh38)}
\format{A data frame with 280956 rows and 3 variables:
\describe{
  \item{ensembl_gene_id}{gene ids in Ensembl 76}
  \item{go_id}{a GO id of the corresponding gene in Ensembl 76}
  \item{name_1006}{name of a GO category associated to the gene in Ensembl 76}
}}
\source{
from several queries with biomaRt (interface to \url{http://www.ensembl.org/biomart}) on Ensembl 76
(\url{aug2014.archive.ensembl.org})
which can not be done anymore on that version specifically
}
\usage{
goAnnot.76
}
\description{
Different gene information based on Ensembl 76 (GRCh38)
}
\keyword{datasets}
