% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gtex.cuff}
\alias{gtex.cuff}
\title{Gene expression for GTEx data (phs000424.v4.p1) aligned with TopHat 2 to Ensembl 76 (GRCh38)
and counted and normalised with cufflinks 2 and then aggregated by tissues/organs.}
\format{A data frame with 53037 genes (ie rows/observations) for 47 tissues (ie columns/variables)
\describe{
 \item{Liver}{numeric, FPKM}
 \item{Kidney}{numeric, FPKM}
 \item{C Trans Fibroblasts}{numeric, FPKM}
 \item{Adrenal}{numeric, FPKM}
 \item{Coronary}{numeric, FPKM}
 \item{Oesophagus}{numeric, FPKM}
 \item{Testis}{numeric, FPKM}
 \item{Stomach}{numeric, FPKM}
 \item{Ovary}{numeric, FPKM}
 \item{Uterus}{numeric, FPKM}
 \item{Aorta}{numeric, FPKM}
 \item{Spleen}{numeric, FPKM}
 \item{Urinarybladder}{numeric, FPKM}
 \item{Colon}{numeric, FPKM}
 \item{Pancreas}{numeric, FPKM}
 \item{Prostate}{numeric, FPKM}
 \item{A Tibial}{numeric, FPKM}
 \item{C EBV Trans Lymph}{numeric, FPKM}
 \item{CML}{numeric, FPKM}
 \item{Adipose}{numeric, FPKM}
 \item{Breast}{numeric, FPKM}
 \item{Pituitary}{numeric, FPKM}
 \item{Vagina}{numeric, FPKM}
 \item{Fallopian tube}{numeric, FPKM}
 \item{Skin}{numeric, FPKM}
 \item{Heart}{numeric, FPKM}
 \item{Skeletal muscle}{numeric, FPKM}
 \item{Nucleus accumbens}{numeric, FPKM}
 \item{Lung}{numeric, FPKM}
 \item{Frontalcortex}{numeric, FPKM}
 \item{Whole blood}{numeric, FPKM}
 \item{Ant cingulate cortex}{numeric, FPKM}
 \item{Cervix}{numeric, FFPKM}
 \item{Nerve tibial}{numeric, FPKM}
 \item{Putamen}{numeric, FPKM}
 \item{Cerebellar Hemi}{numeric, FPKM}
 \item{Thyroid}{numeric, FPKM}
 \item{Small intestine}{numeric, FPKM}
 \item{Salivary gland}{numeric, FPKM}
 \item{Hyppocampus}{numeric, FPKM}
 \item{Amygdala}{numeric, FPKM}
 \item{Cerebellum}{numeric, FFPKM}
 \item{Hypothalamus}{numeric, FPKM}
 \item{Caudate}{numeric, FPKM}
 \item{Spinal cord}{numeric, FPKM}
 \item{Cortex}{numeric, FPKM}
 \item{Substancia nigra}{numeric, FPKM}
}}
\usage{
gtex.cuff
}
\description{
Gene expression for GTEx data (phs000424.v4.p1) aligned with TopHat 2 to Ensembl 76 (GRCh38)
and counted and normalised with cufflinks 2 and then aggregated by tissues/organs.
}
\keyword{datasets}
