% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gtex.cufflinks.tpm.pooled}
\alias{gtex.cufflinks.tpm.pooled}
\title{Gene expression for GTEx (phs000424.v4.p1) mapped to Ensembl 76 (GRCh38)
and quantified with cufflinks, normalised to FPKM before being converted to TPM.
and then all the biological replicates have been aggregated based on their
median value of expression.
See for more details chapters 1 and 2 of Barzine, M.P PhD thesis:
Investigating Normal Human Gene Expression in Tissues
with High-throughput Transcriptomic and Proteomic data.}
\format{A data frame with 57002 transcritps (ie rows/observations) for 47 tissues (ie columns/variables)
\describe{
\item{Liver}{numeric, TPM}
\item{Kidney}{numeric, TPM}
\item{C Trans Fibroblasts}{numeric, TPM}
\item{Adrenal}{numeric, TPM}
\item{Coronary}{numeric, TPM}
\item{Oesophagus}{numeric, TPM}
\item{Testis}{numeric, TPM}
\item{Stomach}{numeric, TPM}
\item{Ovary}{numeric, TPM}
\item{Uterus}{numeric, TPM}
\item{Aorta}{numeric, TPM}
\item{Spleen}{numeric, TPM}
\item{Urinarybladder}{numeric, TPM}
\item{Colon}{numeric, TPM}
\item{Pancreas}{numeric, TPM}
\item{Prostate}{numeric, TPM}
\item{A Tibial}{numeric, TPM}
\item{C EBV Trans Lymph}{numeric, TPM}
\item{CML}{numeric, TPM}
\item{Adipose}{numeric, TPM}
\item{Breast}{numeric, TPM}
\item{Pituitary}{numeric, TPM}
\item{Vagina}{numeric, TPM}
\item{Fallopian tube}{numeric, TPM}
\item{Skin}{numeric, TPM}
\item{Heart}{numeric, TPM}
\item{Skeletal muscle}{numeric, TPM}
\item{Nucleus accumbens}{numeric, TPM}
\item{Lung}{numeric, TPM}
\item{Frontalcortex}{numeric, TPM}
\item{Whole blood}{numeric, TPM}
\item{Ant cingulate cortex}{numeric, TPM}
\item{Cervix}{numeric, TPM}
\item{Nerve tibial}{numeric, TPM}
\item{Putamen}{numeric, TPM}
\item{Cerebellar Hemi}{numeric, TPM}
\item{Thyroid}{numeric, TPM}
\item{Small intestine}{numeric, TPM}
\item{Salivary gland}{numeric, TPM}
\item{Hyppocampus}{numeric, TPM}
\item{Amygdala}{numeric, TPM}
\item{Cerebellum}{numeric, TPM}
\item{Hypothalamus}{numeric, TPM}
\item{Caudate}{numeric, TPM}
\item{Spinalcord}{numeric, TPM}
\item{Cortex}{numeric, TPM}
\item{Substancia nigra}{numeric, TPM}
}}
\usage{
gtex.cufflinks.tpm.pooled
}
\description{
Gene expression for GTEx (phs000424.v4.p1) mapped to Ensembl 76 (GRCh38)
and quantified with cufflinks, normalised to FPKM before being converted to TPM.
and then all the biological replicates have been aggregated based on their
median value of expression.
See for more details chapters 1 and 2 of Barzine, M.P PhD thesis:
Investigating Normal Human Gene Expression in Tissues
with High-throughput Transcriptomic and Proteomic data.
}
\keyword{datasets}
