% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gtex.htseq.fpkm.pooled}
\alias{gtex.htseq.fpkm.pooled}
\title{Gene expression for GTEx (phs000424.v4.p1) mapped to Ensembl 76 (GRCh38)
and quantified with HTSeq-count and normalised with the FPKM method
and then all the biological replicates have been aggregated based on their
median value of expression.
See for more details chapters 1 and 2 of Barzine, M.P PhD thesis:
Investigating Normal Human Gene Expression in Tissues
with High-throughput Transcriptomic and Proteomic data.}
\format{A data frame with 57002 transcritps (ie rows/observations) for 47 tissues (ie columns/variables)
\describe{
\item{Liver}{numeric, FPKM}
\item{Kidney}{numeric, FPKM}
\item{C Trans Fibroblasts}{numeric, FPKM}
\item{Adrenal}{numeric, FPKM}
\item{Coronary}{numeric, FPKM}
\item{Oesophagus}{numeric, FPKM}
\item{Testis}{numeric, FPKM}
\item{Stomach}{numeric, FPKM}
\item{Ovary}{numeric, FPKM}
\item{Uterus}{numeric, FPKM}
\item{Aorta}{numeric, FPKM}
\item{Spleen}{numeric, FPKM}
\item{Urinarybladder}{numeric, FPKM}
\item{Colon}{numeric, FPKM}
\item{Pancreas}{numeric, FPKM}
\item{Prostate}{numeric, FPKM}
\item{A Tibial}{numeric, FPKM}
\item{C EBV Trans Lymph}{numeric, FPKM}
\item{CML}{numeric, FPKM}
\item{Adipose}{numeric, FPKM}
\item{Breast}{numeric, FPKM}
\item{Pituitary}{numeric, FPKM}
\item{Vagina}{numeric, FPKM}
\item{Fallopian tube}{numeric, FPKM}
\item{Skin}{numeric, FPKM}
\item{Heart}{numeric, FPKM}
\item{Skeletal muscle}{numeric, FPKM}
\item{Nucleus accumbens}{numeric, FPKM}
\item{Lung}{numeric, FPKM}
\item{Frontalcortex}{numeric, FPKM}
\item{Whole blood}{numeric, FPKM}
\item{Ant cingulate cortex}{numeric, FPKM}
\item{Cervix}{numeric, FPKM}
\item{Nerve tibial}{numeric, FPKM}
\item{Putamen}{numeric, FPKM}
\item{Cerebellar Hemi}{numeric, FPKM}
\item{Thyroid}{numeric, FPKM}
\item{Small intestine}{numeric, FPKM}
\item{Salivary gland}{numeric, FPKM}
\item{Hyppocampus}{numeric, FPKM}
\item{Amygdala}{numeric, FPKM}
\item{Cerebellum}{numeric, FPKM}
\item{Hypothalamus}{numeric, FPKM}
\item{Caudate}{numeric, FPKM}
\item{Spinalcord}{numeric, FPKM}
\item{Cortex}{numeric, FPKM}
\item{Substancia nigra}{numeric, FPKM}
}}
\usage{
gtex.htseq.fpkm.pooled
}
\description{
Gene expression for GTEx (phs000424.v4.p1) mapped to Ensembl 76 (GRCh38)
and quantified with HTSeq-count and normalised with the FPKM method
and then all the biological replicates have been aggregated based on their
median value of expression.
See for more details chapters 1 and 2 of Barzine, M.P PhD thesis:
Investigating Normal Human Gene Expression in Tissues
with High-throughput Transcriptomic and Proteomic data.
}
\keyword{datasets}
