% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ibm.htseq.raw}
\alias{ibm.htseq.raw}
\title{Gene expression for Illumina Body Map (ArrayExpress E-MTAB-513) mapped to Ensembl 76 (GRCh38)
and quantified with HTSeq-count.
See for more details chapters 1 and 2 of Barzine, M.P PhD thesis:
Investigating Normal Human Gene Expression in Tissues
with High-throughput Transcriptomic and Proteomic data.}
\format{A data frame with 48267 transcritps (ie rows/observations) for 16 tissues (ie columns/variables)
\describe{
\item{Thyroid}{numeric, count}
\item{Testis}{numeric, count}
\item{Ovary}{numeric, count}
\item{Leukocyte}{numeric, count}
\item{Skeletal muscle}{numeric, count}
\item{Prostate}{numeric, count}
\item{Lymph node}{numeric, count}
\item{Lung}{numeric, count}
\item{Adipose}{numeric, count}
\item{Adrenal}{numeric, count}
\item{Brain}{numeric, count}
\item{Breast}{numeric, count}
\item{Colon}{numeric, count}
\item{Kidney}{numeric, count}
\item{Heart}{numeric, count}
\item{Liver}{numeric, count}
}}
\usage{
ibm.htseq.raw
}
\description{
Gene expression for Illumina Body Map (ArrayExpress E-MTAB-513) mapped to Ensembl 76 (GRCh38)
and quantified with HTSeq-count.
See for more details chapters 1 and 2 of Barzine, M.P PhD thesis:
Investigating Normal Human Gene Expression in Tissues
with High-throughput Transcriptomic and Proteomic data.
}
\keyword{datasets}
