% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kuster.ppkm}
\alias{kuster.ppkm}
\title{Protein expression for Kuster Lab data mapped to Ensembl 76 (GRCh38)
and quantified with the PPKM method explained
in the fifth chapter of Barzine, M.P PhD thesis:
Investigating Normal Human Gene Expression in Tissues
with High-throughput Transcriptomic and Proteomic data.
This dataset is the same of kuster.top3.all where the uterus samples have been pooled}
\format{A data frame with 10734 proteins (ie rows/observations) for 35 tissues and body fluids (ie columns/variables)
\describe{
\item{Adrenal}{numeric, PPKM}
\item{Anus}{numeric, PPKM}
\item{Ascites}{numeric, PPKM}
\item{Heart}{numeric, PPKM}
\item{Cortex}{numeric, PPKM}
\item{Cervix}{numeric, PPKM}
\item{Colon}{numeric, PPKM}
\item{Earwax}{numeric, PPKM}
\item{Oesophagus}{numeric, PPKM}
\item{Gallbladder}{numeric, PPKM}
\item{Hair follicle}{numeric, PPKM}
\item{Kidney}{numeric, PPKM}
\item{Liver}{numeric, PPKM}
\item{Lung}{numeric, PPKM}
\item{Lymph node}{numeric, PPKM}
\item{Milk}{numeric, PPKM}
\item{Nasopharynx}{numeric, PPKM}
\item{Oral cavity}{numeric, PPKM}
\item{Ovary}{numeric, PPKM}
\item{Pancreas}{numeric, PPKM}
\item{Placenta}{numeric, PPKM}
\item{Prostate}{numeric, PPKM}
\item{Rectum}{numeric, PPKM}
\item{Saliva}{numeric, PPKM}
\item{Salivary gland}{numeric, PPKM}
\item{Seminal vesicle}{numeric, PPKM}
\item{Skin}{numeric, PPKM}
\item{Spleen}{numeric, PPKM}
\item{Stomach}{numeric, PPKM}
\item{Testis}{numeric, PPKM}
\item{Thyroid}{numeric, PPKM}
\item{Tonsil}{numeric, PPKM}
\item{Tube}{numeric, PPKM}
\item{Uterus}{numeric, PPKM}
\item{Vulva}{numeric, PPM}
}}
\usage{
kuster.ppkm
}
\description{
Protein expression for Kuster Lab data mapped to Ensembl 76 (GRCh38)
and quantified with the PPKM method explained
in the fifth chapter of Barzine, M.P PhD thesis:
Investigating Normal Human Gene Expression in Tissues
with High-throughput Transcriptomic and Proteomic data.
This dataset is the same of kuster.top3.all where the uterus samples have been pooled
}
\keyword{datasets}
