% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kuster.ppm}
\alias{kuster.ppm}
\title{Protein expression for Kuster Lab data mapped to Ensembl 76 (GRCh38)
and quantified with the PPM method which is essentially the same method
as for the PPKM method but the last normalisation step which follows
the TPM (rnaseq) method instead of the FPKM one.
in the fifth chapter of Barzine, M.P PhD thesis:
Investigating Normal Human Gene Expression in Tissues
with High-throughput Transcriptomic and Proteomic data.
This dataset is the same of kuster.top3.all where the uterus samples have been pooled}
\format{A data frame with 10734 proteins (ie rows/observations) for 35 tissues and body fluids (ie columns/variables)
\describe{
\item{Adrenal}{numeric, PPM}
\item{Anus}{numeric, PPM}
\item{Ascites}{numeric, PPM}
\item{Heart}{numeric, PPM}
\item{Cortex}{numeric, PPM}
\item{Cervix}{numeric, PPM}
\item{Colon}{numeric, PPM}
\item{Earwax}{numeric, PPM}
\item{Oesophagus}{numeric, PPM}
\item{Gallbladder}{numeric, PPM}
\item{Hair follicle}{numeric, PPM}
\item{Kidney}{numeric, PPM}
\item{Liver}{numeric, PPM}
\item{Lung}{numeric, PPM}
\item{Lymph node}{numeric, PPM}
\item{Milk}{numeric, PPM}
\item{Nasopharynx}{numeric, PPM}
\item{Oral cavity}{numeric, PPM}
\item{Ovary}{numeric, PPM}
\item{Pancreas}{numeric, PPM}
\item{Placenta}{numeric, PPM}
\item{Prostate}{numeric, PPM}
\item{Rectum}{numeric, PPM}
\item{Saliva}{numeric, PPM}
\item{Salivary gland}{numeric, PPM}
\item{Seminal vesicle}{numeric, PPM}
\item{Skin}{numeric, PPM}
\item{Spleen}{numeric, PPM}
\item{Stomach}{numeric, PPM}
\item{Testis}{numeric, PPM}
\item{Thyroid}{numeric, PPM}
\item{Tonsil}{numeric, PPM}
\item{Tube}{numeric, PPM}
\item{Uterus}{numeric, PPM}
\item{Vulva}{numeric, PPM}
}}
\usage{
kuster.ppm
}
\description{
Protein expression for Kuster Lab data mapped to Ensembl 76 (GRCh38)
and quantified with the PPM method which is essentially the same method
as for the PPKM method but the last normalisation step which follows
the TPM (rnaseq) method instead of the FPKM one.
in the fifth chapter of Barzine, M.P PhD thesis:
Investigating Normal Human Gene Expression in Tissues
with High-throughput Transcriptomic and Proteomic data.
This dataset is the same of kuster.top3.all where the uterus samples have been pooled
}
\keyword{datasets}
