% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kuster.top3.all}
\alias{kuster.top3.all}
\title{Protein expression for Kuster Lab data mapped to Ensembl 76 (GRCh38)
and quantified with the Top3 method (PSM) explained
in the second chapter of Barzine, M.P PhD thesis:
Investigating Normal Human Gene Expression in Tissues
with High-throughput Transcriptomic and Proteomic data.}
\format{A data frame with 5717 proteins (ie rows/observations) for 36 tissues and body fluids (ie columns/variables)
\describe{
\item{Adrenal}{numeric,normalised PSM}
\item{Anus}{numeric,normalised PSM}
\item{Ascites}{numeric,normalised PSM}
\item{Heart}{numeric,normalised PSM}
\item{Cortex}{numeric,normalised PSM}
\item{Cervix}{numeric,normalised PSM}
\item{Colon}{numeric,normalised PSM}
\item{Earwax}{numeric,normalised PSM}
\item{Oesophagus}{numeric,normalised PSM}
\item{Gallbladder}{numeric,normalised PSM}
\item{Hair follicle}{numeric,normalised PSM}
\item{Kidney}{numeric,normalised PSM}
\item{Liver}{numeric,normalised PSM}
\item{Lung}{numeric,normalised PSM}
\item{Lymph node}{numeric,normalised PSM}
\item{Milk}{numeric,normalised PSM}
\item{Nasopharynx}{numeric,normalised PSM}
\item{Oral cavity}{numeric,normalised PSM}
\item{Ovary}{numeric,normalised PSM}
\item{Pancreas}{numeric,normalised PSM}
\item{Placenta}{numeric,normalised PSM}
\item{Prostate}{numeric,normalised PSM}
\item{Rectum}{numeric,normalised PSM}
\item{Saliva}{numeric,normalised PSM}
\item{Salivary gland}{numeric,normalised PSM}
\item{Seminal vesicle}{numeric,normalised PSM}
\item{Skin}{numeric,normalised PSM}
\item{Spleen}{numeric,normalised PSM}
\item{Stomach}{numeric,normalised PSM}
\item{Testis}{numeric,normalised PSM}
\item{Thyroid}{numeric,normalised PSM}
\item{Tonsil}{numeric,normalised PSM}
\item{Tube}{numeric,normalised PSM}
\item{Uterus postmenopause}{numeric,normalised PSM}
\item{Uterus premenopause}{numeric,normalised PSM}
\item{Vulva}{numeric,normalised PSM}
}}
\usage{
kuster.top3.all
}
\description{
Protein expression for Kuster Lab data mapped to Ensembl 76 (GRCh38)
and quantified with the Top3 method (PSM) explained
in the second chapter of Barzine, M.P PhD thesis:
Investigating Normal Human Gene Expression in Tissues
with High-throughput Transcriptomic and Proteomic data.
}
\keyword{datasets}
