% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{mergeUterus}
\alias{mergeUterus}
\title{Allow to merge the two tissues samples related to uterus in the Kuster lab data.}
\usage{
mergeUterus(DF)
}
\arguments{
\item{DF}{one of the data.frame for the kuster lab study.}
}
\value{
the same data.frame except that the uterus premenaupose and postmenaupose are pooled together.
}
\description{
Allow to merge the two tissues samples related to uterus in the Kuster lab data.
}
