% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{NAto0}
\alias{NAto0}
\title{Convert all missing values to 0}
\usage{
NAto0(DF, missingVal = NA)
}
\arguments{
\item{DF}{numeric data.frame}

\item{missingVal}{value to change to 0}
}
\value{
a data.frame with no missing value
}
\description{
Convert all missing values to 0
}
