% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pandey.origin}
\alias{pandey.origin}
\title{Protein expression for Pandey et al. mapped to Ensembl 76 (GRCh38)
and quantified with the Top3 method (PSM) explain
in the second chapter of Barzine, M.P PhD thesis:
Investigating Normal Human Gene Expression in Tissues
with High-throughput Transcriptomic and Proteomic data.}
\format{A data frame with 10272 proteins (ie rows/observations) for 64 tissue samples (ie columns/variables)
\describe{
  \item{X.Tissues}{integer, number of different tissues the protein has been detected}
  \item{maxPeptides}{integer, number of peptides that has benn used for identify the protein}
  \item{medianSpectralCount}{numeric, median number of spectral matching the protein across tissues}
  \item{medianAbundance}{numeric, median abundance for the protein across tissues}
  \item{SpectralCount.Adult_Adrenalgland}{numeric}
  \item{SpectralCount.Adult_Bcells}{numeric}
  \item{SpectralCount.Adult_CD4Tcells}{numeric}
  \item{SpectralCount.Adult_CD8Tcells}{numeric}
  \item{SpectralCount.Adult_Colon}{numeric}
  \item{SpectralCount.Adult_Esophagus}{numeric}
  \item{SpectralCount.Adult_Frontalcortex}{numeric}
  \item{SpectralCount.Adult_Gallbladder}{numeric}
  \item{SpectralCount.Adult_Heart}{numeric}
  \item{SpectralCount.Adult_Kidney}{numeric}
  \item{SpectralCount.Adult_Liver}{numeric}
  \item{SpectralCount.Adult_Lung}{numeric}
  \item{SpectralCount.Adult_Monocytes}{numeric}
  \item{SpectralCount.Adult_NKcells}{numeric}
  \item{SpectralCount.Adult_Ovary}{numeric}
  \item{SpectralCount.Adult_Pancreas}{numeric}
  \item{SpectralCount.Adult_Platelets}{numeric}
  \item{SpectralCount.Adult_Prostate}{numeric}
  \item{SpectralCount.Adult_Rectum}{numeric}
  \item{SpectralCount.Adult_Retina}{numeric}
  \item{SpectralCount.Adult_Spinalcord}{numeric}
  \item{SpectralCount.Adult_Testis}{numeric}
  \item{SpectralCount.Adult_Urinarybladder}{numeric}
  \item{SpectralCount.Fetal_Brain}{numeric}
  \item{SpectralCount.Fetal_Gut}{numeric}
  \item{SpectralCount.Fetal_Heart}{numeric}
  \item{SpectralCount.Fetal_Liver}{numeric}
  \item{SpectralCount.Fetal_Ovary}{numeric}
  \item{SpectralCount.Fetal_Placenta}{numeric}
  \item{SpectralCount.Fetal_Testis}{numeric}
  \item{WithInSampleAbundance.Adult_Adrenalgland}{numeric}
  \item{WithInSampleAbundance.Adult_Bcells}{numeric}
  \item{WithInSampleAbundance.Adult_CD4Tcells}{numeric}
  \item{WithInSampleAbundance.Adult_CD8Tcells}{numeric}
  \item{WithInSampleAbundance.Adult_Colon}{numeric}
  \item{WithInSampleAbundance.Adult_Esophagus}{numeric}
  \item{WithInSampleAbundance.Adult_Frontalcortex}{numeric}
  \item{WithInSampleAbundance.Adult_Gallbladder}{numeric}
  \item{WithInSampleAbundance.Adult_Heart}{numeric}
  \item{WithInSampleAbundance.Adult_Kidney}{numeric}
  \item{WithInSampleAbundance.Adult_Liver}{numeric}
  \item{WithInSampleAbundance.Adult_Lung}{numeric}
  \item{WithInSampleAbundance.Adult_Monocytes}{numeric}
  \item{WithInSampleAbundance.Adult_NKcells}{numeric}
  \item{WithInSampleAbundance.Adult_Ovary}{numeric}
  \item{WithInSampleAbundance.Adult_Pancreas}{numeric}
  \item{WithInSampleAbundance.Adult_Platelets}{numeric}
  \item{WithInSampleAbundance.Adult_Prostate}{numeric}
  \item{WithInSampleAbundance.Adult_Rectum}{numeric}
  \item{WithInSampleAbundance.Adult_Retina}{numeric}
  \item{WithInSampleAbundance.Adult_Spinalcord}{numeric}
  \item{WithInSampleAbundance.Adult_Testis}{numeric}
  \item{WithInSampleAbundance.Adult_Urinarybladder}{numeric}
  \item{WithInSampleAbundance.Fetal_Brain}{numeric}
  \item{WithInSampleAbundance.Fetal_Gut}{numeric}
  \item{WithInSampleAbundance.Fetal_Heart}{numeric}
  \item{WithInSampleAbundance.Fetal_Liver}{numeric}
  \item{WithInSampleAbundance.Fetal_Ovary}{numeric}
  \item{WithInSampleAbundance.Fetal_Placenta}{numeric}
  \item{WithInSampleAbundance.Fetal_Testis}{numeric}
}}
\usage{
pandey.origin
}
\description{
Protein expression for Pandey et al. mapped to Ensembl 76 (GRCh38)
and quantified with the Top3 method (PSM) explain
in the second chapter of Barzine, M.P PhD thesis:
Investigating Normal Human Gene Expression in Tissues
with High-throughput Transcriptomic and Proteomic data.
}
\keyword{datasets}
