% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pandey.ppkm.15}
\alias{pandey.ppkm.15}
\title{Protein expression for Pandey Lab data mapped to Ensembl 76 (GRCh38)
and quantified with the PPKM method explained
in the fifth chapter of Barzine, M.P PhD thesis:
Investigating Normal Human Gene Expression in Tissues
with High-throughput Transcriptomic and Proteomic data.}
\format{A data frame with 12990 proteins (ie rows/observations) for 15 tissues (ie columns/variables)
\describe{
\item{Adrenal}{numeric, PPKM}
\item{Colon}{numeric, PPKM}
\item{Oesophagus}{numeric, PPKM}
\item{Gallbladder}{numeric, PPKM}
\item{Heart}{numeric, PPKM}
\item{Kidney}{numeric, PPKM}
\item{Liver}{numeric, PPKM}
\item{Lung}{numeric, PPKM}
\item{Ovary}{numeric, PPKM}
\item{Pancreas}{numeric, PPKM}
\item{Placenta}{numeric, PPKM}
\item{Prostate}{numeric, PPKM}
\item{Rectum}{numeric, PPKM}
\item{Testis}{numeric, PPKM}
\item{Urinarybladder}{numeric, PPKM}
}}
\usage{
pandey.ppkm.15
}
\description{
Protein expression for Pandey Lab data mapped to Ensembl 76 (GRCh38)
and quantified with the PPKM method explained
in the fifth chapter of Barzine, M.P PhD thesis:
Investigating Normal Human Gene Expression in Tissues
with High-throughput Transcriptomic and Proteomic data.
}
\keyword{datasets}
