% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pandey.ppkm}
\alias{pandey.ppkm}
\title{Protein expression for Pandey et al. mapped to Ensembl 76 (GRCh38)
and quantified with the PPKM method explain
in the fifth chapter of Barzine, M.P PhD thesis:
Investigating Normal Human Gene Expression in Tissues
with High-throughput Transcriptomic and Proteomic data.}
\format{A data frame with 12990 proteins (ie rows/observations) for 15 tissues (ie columns/variables)
\describe{
  \item{Adrenal}{numeric, PPKM}
  \item{Colon}{numeric, PPKM}
  \item{Oesophagus}{numeric, PPKM}
  \item{Gall bladder}{numeric, PPKM}
  \item{Heart}{numeric, PPKM}
  \item{Kidney}{numeric, PPKM}
  \item{Liver}{numeric, PPKM}
  \item{Lung}{numeric, PPKM}
  \item{Ovary}{numeric, PPKM}
  \item{Pancreas}{numeric, PPKM}
  \item{Placenta}{numeric, PPKM}
  \item{Prostate}{numeric, PPKM}
  \item{Rectum}{numeric, PPKM}
  \item{Testis}{numeric, PPKM}
  \item{Urinarybladder}{numeric, PPKM}
}}
\usage{
pandey.ppkm
}
\description{
Protein expression for Pandey et al. mapped to Ensembl 76 (GRCh38)
and quantified with the PPKM method explain
in the fifth chapter of Barzine, M.P PhD thesis:
Investigating Normal Human Gene Expression in Tissues
with High-throughput Transcriptomic and Proteomic data.
}
\keyword{datasets}
