% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pandey.ppm}
\alias{pandey.ppm}
\title{Protein expression for Pandey Lab data mapped to Ensembl 76 (GRCh38)
and quantified with the PPM method which is essentially the same method
as for the PPKM method but the last normalisation step which follows
the TPM (rnaseq) method instead of the FPKM one.
in the fifth chapter of Barzine, M.P PhD thesis:
Investigating Normal Human Gene Expression in Tissues
with High-throughput Transcriptomic and Proteomic data.}
\format{A data frame with 15245 proteins (ie rows/observations) for 30 tissues and cells (ie columns/variables)
\describe{
  \item{Adrenal}{numeric, PPM}
  \item{B cells}{numeric, PPM}
  \item{CD4T cells}{numeric, PPM}
  \item{CD8T cells}{numeric, PPM}
  \item{Colon}{numeric, PPM}
  \item{Oesophagus}{numeric, PPM}
  \item{Fetal Brain}{numeric, PPM}
  \item{Fetal Gut}{numeric, PPM}
  \item{Fetal Heart}{numeric, PPM}
  \item{Fetal Liver}{numeric, PPM}
  \item{Fetal Ovary}{numeric, PPM}
  \item{Placenta}{numeric, PPM}
  \item{Fetal Testis}{numeric, PPM}
  \item{Frontalcortex}{numeric, PPM}
  \item{Gallbladder}{numeric, PPM}
  \item{Heart}{numeric, PPM}
  \item{Kidney}{numeric, PPM}
  \item{Liver}{numeric, PPM}
  \item{Lung}{numeric, PPM}
  \item{Monocytes}{numeric, PPM}
  \item{NK cells}{numeric, PPM}
  \item{Ovary}{numeric, PPM}
  \item{Pancreas}{numeric, PPM}
  \item{Platelets}{numeric, PPM}
  \item{Prostate}{numeric, PPM}
  \item{Rectum}{numeric, PPM}
  \item{Retina}{numeric, PPM}
  \item{Spinalcord}{numeric, PPM}
  \item{Testis}{numeric, PPM}
  \item{Urinarybladder}{numeric, PPM}
}}
\usage{
pandey.ppm
}
\description{
Protein expression for Pandey Lab data mapped to Ensembl 76 (GRCh38)
and quantified with the PPM method which is essentially the same method
as for the PPKM method but the last normalisation step which follows
the TPM (rnaseq) method instead of the FPKM one.
in the fifth chapter of Barzine, M.P PhD thesis:
Investigating Normal Human Gene Expression in Tissues
with High-throughput Transcriptomic and Proteomic data.
}
\keyword{datasets}
