% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pandey.top3.15}
\alias{pandey.top3.15}
\title{Protein expression for Pandey Lab data mapped to Ensembl 76 (GRCh38)
and quantified with the Top3 method (PSM) explained
in the second chapter of Barzine, M.P PhD thesis:
Investigating Normal Human Gene Expression in Tissues
with High-throughput Transcriptomic and Proteomic data.
This is the subset of pandey.top3 which includes the tissues that are shared with Uhlen et al. only.}
\format{A data frame with 6436 proteins (ie rows/observations) for 15 tissue samples (ie columns/variables)
\describe{
\item{Adrenal}{numeric,normalised PSM}
\item{Colon}{numeric, normalised PSM}
\item{Oesophagus}{numeric, normalised PSM}
\item{Gallbladder}{numeric, normalised PSM}
\item{Heart}{numeric, normalised PSM}
\item{Kidney}{numeric, normalised PSM}
\item{Liver}{numeric, normalised PSM}
\item{Lung}{numeric, normalised PSM}
\item{Ovary}{numeric, normalised PSM}
\item{Pancreas}{numeric, normalised PSM}
\item{Placenta}{numeric, normalised PSM}
\item{Prostate}{numeric, normalised PSM}
\item{Rectum}{numeric, normalised PSM}
\item{Testis}{numeric, normalised PSM}
\item{Urinarybladder}{numeric, normalised PSM}
}}
\usage{
pandey.top3.15
}
\description{
Protein expression for Pandey Lab data mapped to Ensembl 76 (GRCh38)
and quantified with the Top3 method (PSM) explained
in the second chapter of Barzine, M.P PhD thesis:
Investigating Normal Human Gene Expression in Tissues
with High-throughput Transcriptomic and Proteomic data.
This is the subset of pandey.top3 which includes the tissues that are shared with Uhlen et al. only.
}
\keyword{datasets}
