% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{selectTigerGenes}
\alias{selectTigerGenes}
\title{TiGER reports TS genes for thirty independent tissues (based on ESTs experiments).
Liu, X., X. Yu, D. J. Zack, H. Zhu, and J. Qian (2008). ‘TiGER: a database for tissue-specific gene expression and regulation’. BMC Bioinformatics 9, p. 271“)}
\format{A vector of selected ensembl gene ids (picked based on TiGER)}
\usage{
selectTigerGenes
}
\description{
vector of ids selected from the panel (allow to showcase various specificity)
}
\keyword{datasets}
