% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{simpleCap}
\alias{simpleCap}
\title{Capitalise the first character of a string}
\usage{
simpleCap(x)
}
\arguments{
\item{x}{Capitalise the first character of a string}
}
\value{
the vector of characters with the first one capitalised
}
\description{
Capitalise the first character of a string
}
