% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TissueCol_12}
\alias{TissueCol_12}
\title{Colour palette for the twelve common tissues
between Uhlén et al. data, GTEx data and Pandey Lab data
designed with
[https://medialab.github.io/iwanthue/](https://medialab.github.io/iwanthue/)}
\format{A name vector with tissues as names and content the hex code for the colour}
\usage{
TissueCol_12
}
\description{
Colour palette for the twelve common tissues
between Uhlén et al. data, GTEx data and Pandey Lab data
designed with
[https://medialab.github.io/iwanthue/](https://medialab.github.io/iwanthue/)
}
\keyword{datasets}
