% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TissueCol_17}
\alias{TissueCol_17}
\title{Colour palette of the union of the common tissues between Kuster data and Uhlén data
with Kuster Lab and GTEx data.
designed with
[https://medialab.github.io/iwanthue/](https://medialab.github.io/iwanthue/)}
\format{A name vector with tissues as names and content the hex code for the colour}
\usage{
TissueCol_17
}
\description{
Colour palette of the union of the common tissues between Kuster data and Uhlén data
with Kuster Lab and GTEx data.
designed with
[https://medialab.github.io/iwanthue/](https://medialab.github.io/iwanthue/)
}
\keyword{datasets}
