% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TissueCol_23}
\alias{TissueCol_23}
\title{Extended colour palette for tissues for Pandey Lab, Kuster, Uhlén data and Gtex data
designed with
[https://medialab.github.io/iwanthue/](https://medialab.github.io/iwanthue/)}
\format{A name vector with tissues as names and content the hex code for the colour}
\usage{
TissueCol_23
}
\description{
Extended colour palette for tissues for Pandey Lab, Kuster, Uhlén data and Gtex data
designed with
[https://medialab.github.io/iwanthue/](https://medialab.github.io/iwanthue/)
}
\keyword{datasets}
