% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TissueCol_62}
\alias{TissueCol_62}
\title{Colour palette covering most of the tissues that are in Pandey, Uhlén, GTEx, Kuster
with their different spelling
designed with
[https://medialab.github.io/iwanthue/](https://medialab.github.io/iwanthue/)}
\format{A name vector with tissues as names and content the hex code for the colour}
\usage{
TissueCol_62
}
\description{
Colour palette covering most of the tissues that are in Pandey, Uhlén, GTEx, Kuster
with their different spelling
designed with
[https://medialab.github.io/iwanthue/](https://medialab.github.io/iwanthue/)
}
\keyword{datasets}
