% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{uhlen.cuff}
\alias{uhlen.cuff}
\title{Gene expression for Uhlen et al. aligned with TopHat 2 to Ensembl 76 (GRCh38)
and counted and normalised with cufflinks 2 and then aggregated by tissues.}
\format{A data frame with 39903 genes (ie rows/observations) for 32 tissues (ie columns/variables)
\describe{
  \item{Adipose}{numeric, FPKM}
  \item{Adrenal}{numeric, FPKM}
  \item{Appendix}{numeric, FPKM}
  \item{Urinarybladder}{numeric, FPKM}
  \item{Bone marrow}{numeric, FPKM}
  \item{Cortex}{numeric, FPKM}
  \item{Colon}{numeric, FPKM}
  \item{Duodenum}{numeric, FPKM}
  \item{Uterus}{numeric, FPKM}
  \item{Oesophagus}{numeric, FPKM}
  \item{Fallopian tube}{numeric, FPKM}
  \item{Gall bladder}{numeric, FPKM}
  \item{Heart}{numeric, FPKM}
  \item{Kidney}{numeric, FPKM}
  \item{Liver}{numeric, FPKM}
  \item{Lung}{numeric, FPKM}
  \item{Lymph node}{numeric, FPKM}
  \item{Ovary}{numeric, FPKM}
  \item{Pancreas}{numeric, FPKM}
  \item{Placenta}{numeric, FPKM}
  \item{Prostate}{numeric, FPKM}
  \item{Rectum}{numeric, FPKM}
  \item{Salivary gland}{numeric, FPKM}
  \item{Skeletal muscle}{numeric, FPKM}
  \item{Skin}{numeric, FPKM}
  \item{Small intestine}{numeric, FPKM}
  \item{Smooth muscle}{numeric, FPKM}
  \item{Spleen}{numeric, FPKM}
  \item{Stomach}{numeric, FPKM}
  \item{Testis}{numeric, FPKM}
  \item{Thyroid}{numeric, FPKM}
  \item{Tonsil}{numeric, FPKM}
}}
\usage{
uhlen.cuff
}
\description{
Gene expression for Uhlen et al. aligned with TopHat 2 to Ensembl 76 (GRCh38)
and counted and normalised with cufflinks 2 and then aggregated by tissues.
}
\keyword{datasets}
