% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{uhlen.cufflinks.fpkm.pooled}
\alias{uhlen.cufflinks.fpkm.pooled}
\title{Gene expression for Uhlén et al. (ArrayExpress E-MTAB-2836) mapped to Ensembl 76 (GRCh38)
and quantified with Cufflinks2 and normalised with the FPKM method.
All the technical replicates have been aggregated together (mean values)
and then the biological replicates have been aggregated by taking the median values.
See for more details chapters 1 and 2 of Barzine, M.P PhD thesis:
Investigating Normal Human Gene Expression in Tissues
with High-throughput Transcriptomic and Proteomic data.}
\format{A data frame with 46932 transcritps (ie rows/observations) for 32 tissues (ie columns/variables)
\describe{
\item{Adipose}{numeric, FPKM}
\item{Adrenal}{numeric, FPKM}
\item{Appendix}{numeric, FPKM}
\item{Urinarybladder}{numeric, FPKM}
\item{Bone marrow}{numeric, FPKM}
\item{Cortex}{numeric, FPKM}
\item{Colon}{numeric, FPKM}
\item{Duodenum}{numeric, FPKM}
\item{Uterus}{numeric, FPKM}
\item{Oesophagus}{numeric, FPKM}
\item{Fallopian tube}{numeric, FPKM}
\item{Gallbladder}{numeric, FPKM}
\item{Heart}{numeric, FPKM}
\item{Kidney}{numeric, FPKM}
\item{Liver}{numeric, FPKM}
\item{Lung}{numeric, FPKM}
\item{Lymph node}{numeric, FPKM}
\item{Ovary}{numeric, FPKM}
\item{Pancreas}{numeric, FPKM}
\item{Placenta}{numeric, FPKM}
\item{Prostate}{numeric, FPKM}
\item{Rectum}{numeric, FPKM}
\item{Salivary gland}{numeric, FPKM}
\item{Skeletal muscle}{numeric, FPKM}
\item{Skin}{numeric, FPKM}
\item{Small intestine}{numeric, FPKM}
\item{Smooth muscle}{numeric, FPKM}
\item{Spleen}{numeric, FPKM}
\item{Stomach}{numeric, FPKM}
\item{Testis}{numeric, FPKM}
\item{Thyroid}{numeric, FPKM}
\item{Tonsil}{numeric, FPKM}
}}
\usage{
uhlen.cufflinks.fpkm.pooled
}
\description{
Gene expression for Uhlén et al. (ArrayExpress E-MTAB-2836) mapped to Ensembl 76 (GRCh38)
and quantified with Cufflinks2 and normalised with the FPKM method.
All the technical replicates have been aggregated together (mean values)
and then the biological replicates have been aggregated by taking the median values.
See for more details chapters 1 and 2 of Barzine, M.P PhD thesis:
Investigating Normal Human Gene Expression in Tissues
with High-throughput Transcriptomic and Proteomic data.
}
\keyword{datasets}
