% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{uhlen.cufflinks.fpkm.pooled.tech}
\alias{uhlen.cufflinks.fpkm.pooled.tech}
\title{Gene expression for Uhlén et al. (ArrayExpress E-MTAB-2836) mapped to Ensembl 76 (GRCh38)
and quantified with Cufflinks2 and normalised with the FPKM method.
All the technical replicates have been aggregated together (mean values).
See for more details chapters 1 and 2 of Barzine, M.P PhD thesis:
Investigating Normal Human Gene Expression in Tissues
with High-throughput Transcriptomic and Proteomic data.}
\format{A data frame with 46932 transcritps (ie rows/observations) for 122 tissues (ie columns/variables)
\describe{#' \item{Adipose.V1}{numeric, FPKM}
\item{Adipose.V20}{numeric, FPKM}
\item{Adipose.V2}{numeric, FPKM}
\item{Adipose.fat_8a}{numeric, FPKM}
\item{Adipose.fat_8b}{numeric, FPKM}
\item{Adrenal.V119}{numeric, FPKM}
\item{Adrenal.V121}{numeric, FPKM}
\item{Adrenal.V122}{numeric, FPKM}
\item{Appendix.V155}{numeric, FPKM}
\item{Appendix.V154}{numeric, FPKM}
\item{Appendix.V160}{numeric, FPKM}
\item{Urinarybladder.V177}{numeric, FPKM}
\item{Urinarybladder.V176}{numeric, FPKM}
\item{Bone.marrow.V250}{numeric, FPKM}
\item{Bone.marrow.V248}{numeric, FPKM}
\item{Bone.marrow.V249}{numeric, FPKM}
\item{Bone.marrow.V230}{numeric, FPKM}
\item{Cerebral.cortex.V29}{numeric, FPKM}
\item{Cerebral.cortex.V103}{numeric, FPKM}
\item{Cerebral.cortex.V102}{numeric, FPKM}
\item{Colon.V10}{numeric, FPKM}
\item{Colon.V15}{numeric, FPKM}
\item{Colon.V11}{numeric, FPKM}
\item{Colon.V22}{numeric, FPKM}
\item{Colon.V14}{numeric, FPKM}
\item{Colon.colon8a}{numeric, FPKM}
\item{Colon.colon8b}{numeric, FPKM}
\item{Duodenum.V150}{numeric, FPKM}
\item{Duodenum.V145}{numeric, FPKM}
\item{Endometrium.V200}{numeric, FPKM}
\item{Endometrium.V165}{numeric, FPKM}
\item{Endometrium.V143}{numeric, FPKM}
\item{Endometrium.endometrium8a}{numeric, FPKM}
\item{Endometrium.endometrium8b}{numeric, FPKM}
\item{Esophagus.V184}{numeric, FPKM}
\item{Esophagus.V185}{numeric, FPKM}
\item{Esophagus.V183}{numeric, FPKM}
\item{Fallopian.tube.fallopiantube_5a}{numeric, FPKM}
\item{Fallopian.tube.fallopiantube_8b}{numeric, FPKM}
\item{Fallopian.tube.fallopiantube_8c}{numeric, FPKM}
\item{Fallopian.tube.fallopiantube_8d}{numeric, FPKM}
\item{Fallopian.tube.fallopiantube_8e}{numeric, FPKM}
\item{Gallbladder.V182}{numeric, FPKM}
\item{Gallbladder.V179}{numeric, FPKM}
\item{Gallbladder.V186}{numeric, FPKM}
\item{Heart.V191}{numeric, FPKM}
\item{Heart.V237}{numeric, FPKM}
\item{Heart.V235}{numeric, FPKM}
\item{Heart.V195}{numeric, FPKM}
\item{Kidney.V24}{numeric, FPKM}
\item{Kidney.V6}{numeric, FPKM}
\item{Kidney.V23}{numeric, FPKM}
\item{Kidney.V5}{numeric, FPKM}
\item{Liver.V110}{numeric, FPKM}
\item{Liver.V111}{numeric, FPKM}
\item{Liver.V108}{numeric, FPKM}
\item{Lung.V130}{numeric, FPKM}
\item{Lung.V81}{numeric, FPKM}
\item{Lung.V80}{numeric, FPKM}
\item{Lung.V133}{numeric, FPKM}
\item{Lung.V131}{numeric, FPKM}
\item{Lymph.node.V193}{numeric, FPKM}
\item{Lymph.node.V164}{numeric, FPKM}
\item{Lymph.node.V157}{numeric, FPKM}
\item{Lymph.node.V192}{numeric, FPKM}
\item{Lymph.node.V190}{numeric, FPKM}
\item{Ovary.V233}{numeric, FPKM}
\item{Ovary.V234}{numeric, FPKM}
\item{Ovary.ovary_8a}{numeric, FPKM}
\item{Pancreas.V232}{numeric, FPKM}
\item{Pancreas.V229}{numeric, FPKM}
\item{Placenta.V224}{numeric, FPKM}
\item{Placenta.V221}{numeric, FPKM}
\item{Placenta.V76}{numeric, FPKM}
\item{Placenta.V223}{numeric, FPKM}
\item{Prostate.V127}{numeric, FPKM}
\item{Prostate.V128}{numeric, FPKM}
\item{Prostate.V129}{numeric, FPKM}
\item{Prostate.V12}{numeric, FPKM}
\item{Rectum.rectum_8a}{numeric, FPKM}
\item{Rectum.rectum_8b}{numeric, FPKM}
\item{Rectum.rectum_8c}{numeric, FPKM}
\item{Rectum.rectum_8d}{numeric, FPKM}
\item{Salivary.gland.V238}{numeric, FPKM}
\item{Salivary.gland.V240}{numeric, FPKM}
\item{Salivary.gland.V239}{numeric, FPKM}
\item{Skeletal.muscle.skeletalmuscle_a}{numeric, FPKM}
\item{Skeletal.muscle.skeletalmuscle_b}{numeric, FPKM}
\item{Skeletal.muscle.skeletalmuscle_c}{numeric, FPKM}
\item{Skeletal.muscle.skeletalmuscle_d}{numeric, FPKM}
\item{Skeletal.muscle.skeletalmuscle_e}{numeric, FPKM}
\item{Skin.V245}{numeric, FPKM}
\item{Skin.V247}{numeric, FPKM}
\item{Skin.V246}{numeric, FPKM}
\item{Small.intestine.V151}{numeric, FPKM}
\item{Small.intestine.V152}{numeric, FPKM}
\item{Small.intestine.V153}{numeric, FPKM}
\item{Small.intestine.V156}{numeric, FPKM}
\item{Smooth.muscle.smoothmuscle_8a}{numeric, FPKM}
\item{Smooth.muscle.smoothmuscle_8b}{numeric, FPKM}
\item{Smooth.muscle.smoothmuscle_8c}{numeric, FPKM}
\item{Spleen.V82}{numeric, FPKM}
\item{Spleen.V83}{numeric, FPKM}
\item{Spleen.V85}{numeric, FPKM}
\item{Spleen.V84}{numeric, FPKM}
\item{Stomach.V91}{numeric, FPKM}
\item{Stomach.V90}{numeric, FPKM}
\item{Stomach.V18}{numeric, FPKM}
\item{Testis.V134}{numeric, FPKM}
\item{Testis.V255}{numeric, FPKM}
\item{Testis.V257}{numeric, FPKM}
\item{Testis.V259}{numeric, FPKM}
\item{Testis.V258}{numeric, FPKM}
\item{Testis.V256}{numeric, FPKM}
\item{Testis.V260}{numeric, FPKM}
\item{Thyroid.V197}{numeric, FPKM}
\item{Thyroid.V196}{numeric, FPKM}
\item{Thyroid.V198}{numeric, FPKM}
\item{Thyroid.V199}{numeric, FPKM}
\item{Tonsil.tonsil_8a1}{numeric, FPKM}
\item{Tonsil.tonsil_8b1}{numeric, FPKM}
\item{Tonsil.tonsil_8e1}{numeric, FPKM}
}}
\usage{
uhlen.cufflinks.fpkm.pooled.tech
}
\description{
Gene expression for Uhlén et al. (ArrayExpress E-MTAB-2836) mapped to Ensembl 76 (GRCh38)
and quantified with Cufflinks2 and normalised with the FPKM method.
All the technical replicates have been aggregated together (mean values).
See for more details chapters 1 and 2 of Barzine, M.P PhD thesis:
Investigating Normal Human Gene Expression in Tissues
with High-throughput Transcriptomic and Proteomic data.
}
\keyword{datasets}
