% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{uhlen.htseq.raw.pooled}
\alias{uhlen.htseq.raw.pooled}
\title{Gene expression for Uhlén et al. (ArrayExpress E-MTAB-2836) mapped to Ensembl 76 (GRCh38)
and quantified with HTSeq-count (not normalised counts).
All the technical replicates have been aggregated together (mean values)
and then the biological replicates have been aggregated by taking the median values.
See for more details chapters 1 and 2 of Barzine, M.P PhD thesis:
Investigating Normal Human Gene Expression in Tissues
with High-throughput Transcriptomic and Proteomic data.}
\format{A data frame with 52655 transcritps (ie rows/observations) for 32 tissues (ie columns/variables)
\describe{
\item{Adipose}{numeric, count}
\item{Adrenal}{numeric, count}
\item{Appendix}{numeric, count}
\item{Urinarybladder}{numeric, count}
\item{Bone marrow}{numeric, count}
\item{Cortex}{numeric, count}
\item{Colon}{numeric, count}
\item{Duodenum}{numeric, count}
\item{Uterus}{numeric, count}
\item{Oesophagus}{numeric, count}
\item{Fallopian tube}{numeric, count}
\item{Gallbladder}{numeric, count}
\item{Heart}{numeric, count}
\item{Kidney}{numeric, count}
\item{Liver}{numeric, count}
\item{Lung}{numeric, count}
\item{Lymph node}{numeric, count}
\item{Ovary}{numeric, count}
\item{Pancreas}{numeric, count}
\item{Placenta}{numeric, count}
\item{Prostate}{numeric, count}
\item{Rectum}{numeric, count}
\item{Salivary gland}{numeric, count}
\item{Skeletal muscle}{numeric, count}
\item{Skin}{numeric, count}
\item{Small intestine}{numeric, count}
\item{Smooth muscle}{numeric, count}
\item{Spleen}{numeric, count}
\item{Stomach}{numeric, count}
\item{Testis}{numeric, count}
\item{Thyroid}{numeric, count}
\item{Tonsil}{numeric, count}
}}
\usage{
uhlen.htseq.raw.pooled
}
\description{
Gene expression for Uhlén et al. (ArrayExpress E-MTAB-2836) mapped to Ensembl 76 (GRCh38)
and quantified with HTSeq-count (not normalised counts).
All the technical replicates have been aggregated together (mean values)
and then the biological replicates have been aggregated by taking the median values.
See for more details chapters 1 and 2 of Barzine, M.P PhD thesis:
Investigating Normal Human Gene Expression in Tissues
with High-throughput Transcriptomic and Proteomic data.
}
\keyword{datasets}
