% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{uhlen.htseq.tpm}
\alias{uhlen.htseq.tpm}
\title{Gene expression for Uhlén et al. (ArrayExpress E-MTAB-2836) mapped to Ensembl 76 (GRCh38)
and quantified with HTSeq-count normalised with the TPM method.
See for more details chapters 1 and 2 of Barzine, M.P PhD thesis:
Investigating Normal Human Gene Expression in Tissues
with High-throughput Transcriptomic and Proteomic data.}
\format{A data frame with 52655 transcritps (ie rows/observations) for 200 samples (ie columns/variables)
\describe{
\item{Salivarygland1}{numeric, TPM}
\item{Lung1}{numeric, TPM}
\item{Liver1}{numeric, TPM}
\item{Heart1}{numeric, TPM}
\item{Lymphnode1}{numeric, TPM}
\item{Prostate1}{numeric, TPM}
\item{Heart2}{numeric, TPM}
\item{Adipose1}{numeric, TPM}
\item{Bonemarrow1}{numeric, TPM}
\item{Bladder1}{numeric, TPM}
\item{Adrenal1}{numeric, TPM}
\item{Placenta1}{numeric, TPM}
\item{Thyroid1}{numeric, TPM}
\item{Spleen1}{numeric, TPM}
\item{Skin1}{numeric, TPM}
\item{Prostate2}{numeric, TPM}
\item{Lung2}{numeric, TPM}
\item{Adipose2}{numeric, TPM}
\item{Adipose3}{numeric, TPM}
\item{Smallintestine1}{numeric, TPM}
\item{Appendix1}{numeric, TPM}
\item{Lung3}{numeric, TPM}
\item{Gallbladder1}{numeric, TPM}
\item{Colon1}{numeric, TPM}
\item{Gallbladder2}{numeric, TPM}
\item{Testis1}{numeric, TPM}
\item{Testis2}{numeric, TPM}
\item{Testis3}{numeric, TPM}
\item{Lung4}{numeric, TPM}
\item{Prostate3}{numeric, TPM}
\item{Bladder2}{numeric, TPM}
\item{Heart3}{numeric, TPM}
\item{Colon2}{numeric, TPM}
\item{Thyroid2}{numeric, TPM}
\item{Prostate4}{numeric, TPM}
\item{Gallbladder3}{numeric, TPM}
\item{Endometrium1}{numeric, TPM}
\item{Esophagus1}{numeric, TPM}
\item{Thyroid3}{numeric, TPM}
\item{Smallintestine2}{numeric, TPM}
\item{Prostate5}{numeric, TPM}
\item{Appendix2}{numeric, TPM}
\item{Heart4}{numeric, TPM}
\item{Endometrium2}{numeric, TPM}
\item{Stomach1}{numeric, TPM}
\item{Bladder3}{numeric, TPM}
\item{Lymphnode2}{numeric, TPM}
\item{Skin2}{numeric, TPM}
\item{Lymphnode3}{numeric, TPM}
\item{Placenta2}{numeric, TPM}
\item{Placenta3}{numeric, TPM}
\item{Skin3}{numeric, TPM}
\item{Placenta4}{numeric, TPM}
\item{Adipose4}{numeric, TPM}
\item{Stomach2}{numeric, TPM}
\item{Ovary1}{numeric, TPM}
\item{Smallintestine3}{numeric, TPM}
\item{Salivarygland2}{numeric, TPM}
\item{Kidney1}{numeric, TPM}
\item{Heart5}{numeric, TPM}
\item{Adrenal2}{numeric, TPM}
\item{Endometrium3}{numeric, TPM}
\item{Lymphnode4}{numeric, TPM}
\item{Smallintestine4}{numeric, TPM}
\item{Heart6}{numeric, TPM}
\item{Lymphnode5}{numeric, TPM}
\item{Testis4}{numeric, TPM}
\item{Adrenal3}{numeric, TPM}
\item{Lymphnode6}{numeric, TPM}
\item{Liver2}{numeric, TPM}
\item{Bonemarrow2}{numeric, TPM}
\item{Bonemarrow3}{numeric, TPM}
\item{Thyroid4}{numeric, TPM}
\item{Esophagus2}{numeric, TPM}
\item{Placenta5}{numeric, TPM}
\item{Colon3}{numeric, TPM}
\item{Skin4}{numeric, TPM}
\item{Ovary2}{numeric, TPM}
\item{Colon4}{numeric, TPM}
\item{Bonemarrow4}{numeric, TPM}
\item{Spleen2}{numeric, TPM}
\item{Bonemarrow5}{numeric, TPM}
\item{Prostate6}{numeric, TPM}
\item{Smallintestine5}{numeric, TPM}
\item{Smallintestine6}{numeric, TPM}
\item{Prostate7}{numeric, TPM}
\item{Esophagus3}{numeric, TPM}
\item{Thyroid5}{numeric, TPM}
\item{Heart7}{numeric, TPM}
\item{Liver3}{numeric, TPM}
\item{Testis5}{numeric, TPM}
\item{Spleen3}{numeric, TPM}
\item{Adrenal4}{numeric, TPM}
\item{Salivarygland3}{numeric, TPM}
\item{Smallintestine7}{numeric, TPM}
\item{Salivarygland4}{numeric, TPM}
\item{Bladder4}{numeric, TPM}
\item{Thyroid6}{numeric, TPM}
\item{Smallintestine8}{numeric, TPM}
\item{Lung5}{numeric, TPM}
\item{Bonemarrow6}{numeric, TPM}
\item{Lymphnode7}{numeric, TPM}
\item{Gallbladder4}{numeric, TPM}
\item{Thyroid7}{numeric, TPM}
\item{Pancreas1}{numeric, TPM}
\item{Heart8}{numeric, TPM}
\item{Adipose5}{numeric, TPM}
\item{Cerebralcortex1}{numeric, TPM}
\item{Endometrium4}{numeric, TPM}
\item{Esophagus4}{numeric, TPM}
\item{Heart9}{numeric, TPM}
\item{Pancreas2}{numeric, TPM}
\item{Appendix3}{numeric, TPM}
\item{Endometrium5}{numeric, TPM}
\item{Lung6}{numeric, TPM}
\item{Lymphnode8}{numeric, TPM}
\item{Lymphnode9}{numeric, TPM}
\item{Duodenum1}{numeric, TPM}
\item{Kidney2}{numeric, TPM}
\item{Lung7}{numeric, TPM}
\item{Duodenum2}{numeric, TPM}
\item{Testis6}{numeric, TPM}
\item{Bladder5}{numeric, TPM}
\item{Spleen4}{numeric, TPM}
\item{Salivarygland5}{numeric, TPM}
\item{Adrenal5}{numeric, TPM}
\item{Liver4}{numeric, TPM}
\item{Adrenal6}{numeric, TPM}
\item{Bladder6}{numeric, TPM}
\item{Appendix4}{numeric, TPM}
\item{Cerebralcortex2}{numeric, TPM}
\item{Testis7}{numeric, TPM}
\item{Duodenum3}{numeric, TPM}
\item{Ovary3}{numeric, TPM}
\item{Salivarygland6}{numeric, TPM}
\item{Skin5}{numeric, TPM}
\item{Duodenum4}{numeric, TPM}
\item{Colon5}{numeric, TPM}
\item{Liver5}{numeric, TPM}
\item{Skin6}{numeric, TPM}
\item{Appendix5}{numeric, TPM}
\item{Pancreas3}{numeric, TPM}
\item{Stomach3}{numeric, TPM}
\item{Kidney3}{numeric, TPM}
\item{Bonemarrow7}{numeric, TPM}
\item{Gallbladder5}{numeric, TPM}
\item{Lymphnode10}{numeric, TPM}
\item{Esophagus5}{numeric, TPM}
\item{Spleen5}{numeric, TPM}
\item{Gallbladder6}{numeric, TPM}
\item{Lymphnode11}{numeric, TPM}
\item{Placenta6}{numeric, TPM}
\item{Cerebralcortex3}{numeric, TPM}
\item{Placenta7}{numeric, TPM}
\item{Pancreas4}{numeric, TPM}
\item{Gallbladder7}{numeric, TPM}
\item{Appendix6}{numeric, TPM}
\item{Ovary4}{numeric, TPM}
\item{Thyroid8}{numeric, TPM}
\item{Colon6}{numeric, TPM}
\item{Stomach4}{numeric, TPM}
\item{Bonemarrow8}{numeric, TPM}
\item{Lung8}{numeric, TPM}
\item{Lymphnode12}{numeric, TPM}
\item{Esophagus6}{numeric, TPM}
\item{Endometrium6}{numeric, TPM}
\item{Thyroid9}{numeric, TPM}
\item{Testis8}{numeric, TPM}
\item{Lymphnode13}{numeric, TPM}
\item{Kidney4}{numeric, TPM}
\item{Endometrium7}{numeric, TPM}
\item{Colon7}{numeric, TPM}
\item{Colon8}{numeric, TPM}
\item{Endometrium8}{numeric, TPM}
\item{Endometrium9}{numeric, TPM}
\item{Fallopiantube1}{numeric, TPM}
\item{Fallopiantube2}{numeric, TPM}
\item{Fallopiantube3}{numeric, TPM}
\item{Fallopiantube4}{numeric, TPM}
\item{Fallopiantube5}{numeric, TPM}
\item{Fallopiantube6}{numeric, TPM}
\item{Adipose6}{numeric, TPM}
\item{Adipose7}{numeric, TPM}
\item{Rectum1}{numeric, TPM}
\item{Rectum2}{numeric, TPM}
\item{Rectum3}{numeric, TPM}
\item{Rectum4}{numeric, TPM}
\item{Skeletalmuscle1}{numeric, TPM}
\item{Skeletalmuscle2}{numeric, TPM}
\item{Skeletalmuscle3}{numeric, TPM}
\item{Skeletalmuscle4}{numeric, TPM}
\item{Skeletalmuscle5}{numeric, TPM}
\item{Skeletalmuscle6}{numeric, TPM}
\item{Smoothmuscle1}{numeric, TPM}
\item{Smoothmuscle2}{numeric, TPM}
\item{Smoothmuscle3}{numeric, TPM}
\item{Tonsil1}{numeric, TPM}
\item{Tonsil2}{numeric, TPM}
\item{Tonsil3}{numeric, TPM}
\item{Ovary5}{numeric, TPM}
}}
\usage{
uhlen.htseq.tpm
}
\description{
Gene expression for Uhlén et al. (ArrayExpress E-MTAB-2836) mapped to Ensembl 76 (GRCh38)
and quantified with HTSeq-count normalised with the TPM method.
See for more details chapters 1 and 2 of Barzine, M.P PhD thesis:
Investigating Normal Human Gene Expression in Tissues
with High-throughput Transcriptomic and Proteomic data.
}
\keyword{datasets}
