% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pandey.top3}
\alias{pandey.top3}
\title{Protein expression for Pandey et al. mapped to Ensembl 76 (GRCh38)
and quantified with the Top3 method (PSM) explain
in the second chapter of Barzine, M.P PhD thesis:
Investigating Normal Human Gene Expression in Tissues
with High-throughput Transcriptomic and Proteomic data.}
\format{A data frame with 6436 proteins (ie rows/observations) for 15 tissue samples (ie columns/variables)
\describe{
  \item{Adrenal}{numeric,PSM}
  \item{Colon}{numeric,PSM}
  \item{Oesophagus}{numeric,PSM}
  \item{Gall bladder}{numeric,PSM}
  \item{Heart}{numeric,PSM}
  \item{Kidney}{numeric,PSM}
  \item{Liver}{numeric,PSM}
  \item{Lung}{numeric,PSM}
  \item{Ovary}{numeric,PSM}
  \item{Pancreas}{numeric,PSM}
  \item{Placenta}{numeric,PSM}
  \item{Prostate}{numeric,PSM}
  \item{Rectum}{numeric,PSM}
  \item{Testis}{numeric,PSM}
  \item{Urinarybladder}{numeric,PSM}
}}
\usage{
pandey.top3
}
\description{
Protein expression for Pandey et al. mapped to Ensembl 76 (GRCh38)
and quantified with the Top3 method (PSM) explain
in the second chapter of Barzine, M.P PhD thesis:
Investigating Normal Human Gene Expression in Tissues
with High-throughput Transcriptomic and Proteomic data.
}
\keyword{datasets}
